makesymlinks: Add directory changing code
By adding this code, it is possible to stay inside one of the packages
to facilitate testing. makesymlinks can be called from any working
directory, and the script will automatically go to the right place to do
its work, as well as pop the user's working directory off the stack.
Then the user can continue working on whatever config they're working
on, if needed