i3/config: Edit lock command
The default behavior of i3lock is to use a white screen, which is rather
annoying. So change that to black.

Additionally, the default behavior is to include feedback for keypresses
and wrong passwords. To further discourage wannabe physical snoopers,
the -u option to disable visual feedback. It might be a bit annoying,
especially if I myself type my password wrong, but that little peace of
mind would be nice.