i3/config: Shuffle Caps Lock functionality
Caps Lock isn't a function that's common enough to warrant its own key.
After all, the functionality is activated by double-tapping Shift on
mobile keyboards. Depressing both Shift keys is close enough to that, so
use it. Then, we can use the Caps Lock key to get extra functionality in
the form of the Compose key