i3/adjust-volume: Tweak notification settings
notif_cmd is declared to DRY out the code, so that if I need to make
further adjustments to the notify-send command, it's only necessary to
change in one place.

For the command itself, the hint clears the notification from the
notification tray immediately. Without it, the notification remains in
the tray indefinitely, which is not necessary or desired. -t sets a
timeout for the notification bubble. For this particular application,
250 milliseconds is sufficient.