top/bash_profile: Set PATH for rofi
rofi wasn't seeing the scripts in ~/bin, even though that directory is
in my PATH. Searching around indicates that the $PATH should be set in
~/.profile, but that is already set, and it still wasn't working. A note
at the top of ~/.profile states that ~/.profile is not sourced if
~/.bash_profile exists. In my case, ~/.bash_profile does exist, and it's
how I choose my GUI environment. Ergo, the PATH needs to be set in this
file for rofi to work correctly.