i3/config: Disable floating windows
Floating windows are a hassle in i3 at times, because the only way to
move focus between floating windows and tiled windows is via the mouse.
This defeats the purpose of a tiling WM. This setting makes some windows
look weird, but so far I'm willing to put up with that.