Update package lists
The lists are shorter, because they only contain packages that were installed
explicitly; they exclude dependencies.

These lists are being updated, because I plan to buy a second SSD for the
laptop, so that I can have a dedicated Windows disk for gaming and a dedicated
Linux disk for everything else, as opposed to the current single-disk dual-boot.

The plan is as follows:
1. Save to my external HDD any Linux folders I need
2. Install the new SSD hardware
3. Install Arch to the new SSD, using the entire space
4. Once the new Arch install is confirmed working, wipe out the Linux partitions
   (/, /home) on the original SSD, *while keeping the existing EFI partition*
5. Use GParted on an Ubuntu ISO to enlarge the Windows partition to fill the
   remainder of the original disk