Move emacs config into ~/.emacs.d/
Emacs 29 has been released, and,it includes out-of-the-box some packages
that I previously had to install explicitly.  So it's as good a time as
any to reconsider the config.  I still have to figure out what's
included and what's not—what needs to be installed and what can be cut,
so the new emacs 29 config is just starting with the basic interface
config that I like, and packages will be added as needed